Renaming and general improvements
