+++ models/stop.py
@@ -4,12 +4,12 @@ from math import sqrt
 import helpers.sheet
 
 from models.match import Match
-from models.service import ServiceGroup
+from models.schedule import Schedule
 
 class Stop:
     '''A location where a vehicle stops along a trip'''
     
-    __slots__ = ('system', 'id', 'number', 'name', 'lat', 'lon', 'departures', 'service_group', 'sheets')
+    __slots__ = ('system', 'id', 'number', 'name', 'lat', 'lon', 'departures', 'schedule', 'sheets')
     
     @classmethod
     def from_csv(cls, row, system, departures):
@@ -31,8 +31,8 @@ class Stop:
         self.departures = departures
         
         services = {d.trip.service for d in departures if d.trip is not None}
-        self.service_group = ServiceGroup.combine(system, services)
-        self.sheets = helpers.sheet.combine(services)
+        self.schedule = Schedule.combine(system, [s.schedule for s in services])
+        self.sheets = helpers.sheet.combine(system, services)
     
     def __str__(self):
         return self.name
@@ -47,14 +47,6 @@ class Stop:
         if self.name == other.name:
             return self.number < other.number
         return self.name < other.name
-    
-    @property
-    def is_current(self):
-        '''Checks if this stop is included in the current sheet'''
-        for departure in self.departures:
-            if departure.is_current:
-                return True
-        return False
     
     @property
     def nearby_stops(self):
