+++ views/pages/route/schedule.tpl
@@ -23,12 +23,12 @@
         Please check again later!
     </p>
 % else:
-    % sheets = [s for s in route.sheets if s.is_current]
-        
+    % sheets = route.sheets
+    
     % if len(sheets) > 1 or (len(sheets) == 1 and len(sheets[0].service_groups) > 1):
         % include('components/sheet_navigation', sheets=sheets)
     % end
-
+    
     <div class="container">
         % for sheet in sheets:
             % for service_group in sheet.service_groups:
