+++ views/pages/stop/schedule.tpl
@@ -24,7 +24,7 @@
         Please check again later!
     </p>
 % else:
-    % sheets = [s for s in stop.sheets if s.is_current]
+    % sheets = stop.sheets
     
     % if len(sheets) > 1 or (len(sheets) == 1 and len(sheets[0].service_groups) > 1):
         % include('components/sheet_navigation', sheets=sheets)
