+++ views/pages/stops.tpl
@@ -10,13 +10,38 @@
 </div>
 
 % if system is None:
-    <p>
-        Stops can only be viewed for individual systems.
-        Please choose a system.
-    </p>
-    <div class="non-desktop">
-        % include('components/systems')
-    </div>
+    <p>Choose a system to see individual stops.</p>
+    <table class="striped">
+        <thead>
+            <tr>
+                <th>System</th>
+                <th># Stops</th>
+                <th>Service Days</th>
+            </tr>
+        </thead>
+        <tbody>
+            % for region in regions:
+                % region_systems = [s for s in systems if s.region == region]
+                % if len(region_systems) > 0:
+                    <tr class="section">
+                        <td colspan="3">
+                            {{ region }}
+                        </td>
+                    </tr>
+                    <tr class="display-none"></tr>
+                    % for region_system in region_systems:
+                        <tr>
+                            <td><a href="{{ get_url(region_system, path) }}">{{ region_system }}</a></td>
+                            <td>{{ len(region_system.get_stops()) }}</td>
+                            <td>
+                                % include('components/schedule_indicator', schedule=region_system.schedule, compact=True)
+                            </td>
+                        </tr>
+                    % end
+                % end
+            % end
+        </tbody>
+    </table>
 % else:
     % stops = system.get_stops()
     % if len(stops) == 0 and search is None:
